+++ packages/next-common/components/login/evm/form.jsx
@@ -2,18 +2,17 @@ import { useAccount, useConnect } from "wagmi";
 import { useDispatch } from "react-redux";
 import { setConnectPopupView } from "next-common/store/reducers/connectPopupSlice";
 import PrimaryButton from "next-common/lib/button/primary";
-import WalletOption from "../../wallet/walletOption";
 import AddressSelect from "../../addressSelect";
 import { useEffect, useState } from "react";
 import { useWeb3Login } from "next-common/hooks/connect/useWeb3Login";
 import { CONNECT_POPUP_VIEWS } from "next-common/utils/constants";
-import { useEVMWalletOptions } from "next-common/hooks/connect/useEVMWalletOptions";
 import { useConnectedAccountContext } from "next-common/context/connectedAccount";
 import { find } from "lodash-es";
 import { SystemLoading } from "@osn/icons/subsquare";
 import WalletTypes from "next-common/utils/consts/walletTypes";
 import { useAccounts } from "next-common/hooks/connect/evm/useAccounts";
 import ErrorMessage from "next-common/components/styled/errorMessage";
+import EVMWalletOptions from "next-common/components/wallet/evmWalletOptions";
 
 export default function LoginEVMForm() {
   const dispatch = useDispatch();
@@ -49,50 +48,25 @@ export default function LoginEVMForm() {
     });
   }
 
-  const evmOptions = useEVMWalletOptions();
-
-  const walletOptions = evmOptions.map((option) => {
-    let icon;
-    if (option.logo) {
-      icon = <option.logo className="w-6 h-6" />;
-    } else if (option.connector.icon) {
-      icon = <img src={option.connector.icon} className="w-6 h-6" />;
-    }
-
-    const installed = !!option.connector;
-
-    return (
-      <WalletOption
-        key={option.extensionName}
-        installed={installed}
-        selected={selectedConnector?.id === option.connector?.id}
-        onClick={() => {
-          if (!installed) {
-            return;
-          }
-
-          connect(
-            { connector: option.connector },
-            {
-              onSuccess() {
-                setSelectedConnector(option.connector);
-              },
-            },
-          );
-        }}
-        logo={icon}
-        title={option.title}
-      />
-    );
-  });
-
   return (
     <div>
       <div className="text14Bold text-textPrimary mb-2">EVM Wallet</div>
 
       <div className="space-y-6">
         <div className="grid grid-cols-2 gap-2 max-sm:grid-cols-1">
-          {walletOptions}
+          <EVMWalletOptions
+            selectedConnector={selectedConnector}
+            onSelect={(option) => {
+              connect(
+                { connector: option.connector },
+                {
+                  onSuccess() {
+                    setSelectedConnector(option.connector);
+                  },
+                },
+              );
+            }}
+          />
         </div>
 
         {isConnecting && (
